Adapt scss files based on new architecture's norms
