+++ src/pages/SubmitComplain/SubmitComplaitPhoto.tsx
@@ -1,0 +1,20 @@
+import { FiCamera } from 'react-icons/fi';
+import { useHistory } from 'react-router-dom';
+import Button from '../../components/Button';
+
+const SubmitComplaitPhoto = () => {
+	const history = useHistory();
+	const onSubmit = () => {
+		history.push('/submit-complaint/photo', history.location.state);
+	};
+
+	return (
+		<div className='takePhoto'>
+			<div />
+			<FiCamera />
+			<Button text='Continuar' onClick={onSubmit} />
+		</div>
+	);
+};
+
+export default SubmitComplaitPhoto;
