+++ src/styles/layouts/_home.scss
@@ -1,5 +1,3 @@
-@import 'main.scss';
-
 .home {
     display: flex;
     justify-content: center;
