+++ src/styles/layouts/_index.scss
@@ -1,0 +1,6 @@
+@import 'home';
+@import 'submitComplaintGeolocation';
+@import 'submitComplaintPhoto';
+@import 'submitComplaintOptionPhoto';
+@import 'submitComplaintConclude';
+@import 'submitComplaintInfos';
