+++ src/styles/layouts/_submitComplaintInfos.scss
@@ -1,5 +1,3 @@
-@import 'main.scss';
-
 .submitComplaintInfos {
     display: flex;
     flex-direction: column;
@@ -8,9 +6,8 @@
     justify-content: space-around;
     height: calc(100vh - 7.25rem);
 
-    color: $branco;
-    font-family: Spartan;
-    font-size: 18px;
+    color: var(--text-color);
+    @include font-style($style-name: 'default');
 
     &__content {
         display: flex;
